# docs/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Twidder
  contact: {}
host: localhost:8080
basePath: /
securityDefinitions: {}
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /auth/signup:
    post:
      summary: Create User budi
      tags:
        - Auth
      operationId: CreateUserbudi
      deprecated: false
      produces:
        - application/json; charset=utf-8
      consumes:
        - application/json
      parameters:
        - name: Content-Type
          in: header
          required: false
          enum:
            - application/json
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/CreateUserbudiRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/IvanLanin'
          examples:
            application/json; charset=utf-8:
              payload:
                ID: 2
                Username: ivanlanin
                DisplayName: Ivan Lanin
                Email: ivanlanin@gmail.com
                CreatedAt: '2021-05-11T10:45:10.087+07:00'
                UpdatedAt: '2021-05-11T10:45:10.087+07:00'
                DeletedAt: null
          headers:
            Date:
              type: string
              default: Tue, 11 May 2021 03:45:10 GMT
            Content-Length:
              type: string
              default: '205'
  /auth/login:
    post:
      summary: Login budi
      tags:
        - Auth
      operationId: Loginbudi
      deprecated: false
      produces:
        - application/json; charset=utf-8
      consumes:
        - application/json
      parameters:
        - name: Content-Type
          in: header
          required: false
          enum:
            - application/json
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/LoginbudiRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/LoginIvanLanin'
          examples:
            application/json; charset=utf-8:
              code: 200
              expire: '2021-05-11T11:20:39+07:00'
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjA3MDY4MzksImlkIjoxLCJvcmlnX2lhdCI6MTYyMDcwMzIzOX0.7j7ymB8HyhOMA933Xz2AEWXKDPNB5ovdASUVKm6roLQ
          headers:
            Set-Cookie:
              type: string
              default: token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjA3MDY4MzksImlkIjoxLCJvcmlnX2lhdCI6MTYyMDcwMzIzOX0.7j7ymB8HyhOMA933Xz2AEWXKDPNB5ovdASUVKm6roLQ; Path=/; Max-Age=3600; HttpOnly
            Date:
              type: string
              default: Tue, 11 May 2021 03:20:39 GMT
            Content-Length:
              type: string
              default: '204'
  /auth/logout:
    post:
      summary: Logout
      tags:
        - Auth
      operationId: Logout
      deprecated: false
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: Content-Type
          in: header
          required: false
          enum:
            - application/json
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/LogoutRequest'
      responses:
        '200':
          description: ''
          headers: {}
...
  /auth/reactivate-account:
    post:
      summary: Reactivate Account
      tags:
        - Auth
      operationId: ReactivateAccount
      deprecated: false
      produces:
        - application/json; charset=utf-8
      consumes:
        - application/json
      parameters:
        - name: Content-Type
          in: header
          required: false
          enum:
            - application/json
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/ReactivateAccountRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/LoginIhza'
          examples:
            application/json; charset=utf-8:
              code: 200
              expire: '2021-05-11T11:20:39+07:00'
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjA3MDY4MzksImlkIjoxLCJvcmlnX2lhdCI6MTYyMDcwMzIzOX0.7j7ymB8HyhOMA933Xz2AEWXKDPNB5ovdASUVKm6roLQ
          headers:
            Set-Cookie:
              type: string
              default: token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjA3MDY4MzksImlkIjoxLCJvcmlnX2lhdCI6MTYyMDcwMzIzOX0.7j7ymB8HyhOMA933Xz2AEWXKDPNB5ovdASUVKm6roLQ; Path=/; Max-Age=3600; HttpOnly
            Date:
              type: string
              default: Tue, 11 May 2021 03:20:39 GMT
            Content-Length:
              type: string
              default: '204'
...
definitions:
  CreateUserbudiRequest:
    title: CreateUserbudiRequest
    example:
      email: budi@yahoo.com
      userName: budi
      password: ArifBudiman123
      displayName: Arif Budiman
    type: object
    properties:
      email:
        type: string
      userName:
        type: string
      password:
        type: string
      displayName:
        type: string
    required:
      - email
      - userName
      - password
      - displayName
  IvanLanin:
    title: IvanLanin
    example:
      payload:
        ID: 2
        Username: ivanlanin
        DisplayName: Ivan Lanin
        Email: ivanlanin@gmail.com
        CreatedAt: '2021-05-11T10:45:10.087+07:00'
        UpdatedAt: '2021-05-11T10:45:10.087+07:00'
        DeletedAt: null
    type: object
    properties:
      payload:
        $ref: '#/definitions/Payload'
    required:
      - payload
  Payload:
    title: Payload
    example:
      ID: 2
      Username: ivanlanin
      DisplayName: Ivan Lanin
      Email: ivanlanin@gmail.com
      CreatedAt: '2021-05-11T10:45:10.087+07:00'
      UpdatedAt: '2021-05-11T10:45:10.087+07:00'
      DeletedAt: null
    type: object
    properties:
      ID:
        type: integer
        format: int32
      Username:
        type: string
      DisplayName:
        type: string
      Email:
        type: string
      CreatedAt:
        type: string
      UpdatedAt:
        type: string
      DeletedAt:
        type: string
    required:
      - ID
      - Username
      - DisplayName
      - Email
      - CreatedAt
      - UpdatedAt
      - DeletedAt
...
  LoginbudiRequest:
    title: LoginbudiRequest
    example:
      username: budi
      password: ArifBudiman123
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    required:
      - username
      - password
  LoginIvanLanin:
    title: LoginIvanLanin
    example:
      code: 200
      expire: '2021-05-11T11:20:39+07:00'
      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjA3MDY4MzksImlkIjoxLCJvcmlnX2lhdCI6MTYyMDcwMzIzOX0.7j7ymB8HyhOMA933Xz2AEWXKDPNB5ovdASUVKm6roLQ
    type: object
    properties:
      code:
        type: integer
        format: int32
      expire:
        type: string
      token:
        type: string
    required:
      - code
      - expire
      - token
  LoginIhza:
    title: LoginIhza
    example:
      code: 200
      expire: '2021-05-11T11:20:39+07:00'
      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjA3MDY4MzksImlkIjoxLCJvcmlnX2lhdCI6MTYyMDcwMzIzOX0.7j7ymB8HyhOMA933Xz2AEWXKDPNB5ovdASUVKm6roLQ
    type: object
    properties:
      code:
        type: integer
        format: int32
      expire:
        type: string
      token:
        type: string
    required:
      - code
      - expire
      - token
...
  LogoutRequest:
    title: LogoutRequest
    example:
      username: Ihzam31415
      password: IhzaKeren99
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    required:
      - username
      - password
  ReactivateAccountRequest:
    title: ReactivateAccountRequest
    example:
      username: ivanlanin
      password: IvanLaninKeren123
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    required:
      - username
      - password
...